refactor: extract normalizeIdGroups to its own dep-free module
The vitest test only uses normalizeIdGroups, a pure function with no
third-party deps. Keeping it in promptExtractor.js forced vitest to
resolve @babel/parser (only in tools/node_modules) from root, causing
the CI failure fixed in the previous commit.

Extract normalizeIdGroups + its .d.ts to tools/normalizeIdGroups.js so
the test imports a dep-free module. Any future heavy imports added to
promptExtractor.js won't affect vitest. Remove @babel/parser from root
devDependencies (no longer needed).

promptExtractor.js re-exports normalizeIdGroups for callers that import
it from the old path.
