+++ package.json
@@ -64,7 +64,6 @@
   },
   "homepage": "https://github.com/skrabe/tweakcc-fixed#readme",
   "devDependencies": {
-    "@babel/parser": "^7.28.5",
     "@eslint/js": "^9.33.0",
     "@types/node": "^20.19.10",
     "@types/react": "^19.2.14",
