+++ tools/promptExtractor.d.ts
@@ -1,22 +1,12 @@
-export interface PromptEntry {
-  id: string;
-  name?: string;
-  description?: string;
-  pieces: string[];
-  identifiers: (number | string)[];
-  identifierMap: Record<string, string>;
-  version?: string;
-  start: number;
-  end: number;
-}
+import type { PromptEntry } from './normalizeIdGroups.js';
+
+export type { PromptEntry };
 
 export interface ExtractResult {
   prompts: PromptEntry[];
 }
 
-export declare function normalizeIdGroups(
-  prompts: PromptEntry[]
-): PromptEntry[];
+export { normalizeIdGroups } from './normalizeIdGroups.js';
 export default function extractStrings(
   filepath: string,
   minLength?: number
